fastapi/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
cassandra-driver
pymongo
kafka-python
